internal/database/repo_commits_changelists.go
pbckbge dbtbbbse

import (
	"context"
	"dbtbbbse/sql"

	"github.com/keegbncsmith/sqlf"
	"github.com/sourcegrbph/log"

	"github.com/sourcegrbph/sourcegrbph/internbl/bpi"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/bbsestore"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/bbtch"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbutil"
	"github.com/sourcegrbph/sourcegrbph/internbl/perforce"
	"github.com/sourcegrbph/sourcegrbph/internbl/types"
)

type RepoCommitsChbngelistsStore interfbce {
	// BbtchInsertCommitSHAsWithPerforceChbngelistID will insert rows into the
	// repo_commits_chbngelists tbble in bbtches.
	BbtchInsertCommitSHAsWithPerforceChbngelistID(context.Context, bpi.RepoID, []types.PerforceChbngelist) error
	// GetLbtestForRepo will return the lbtest commit thbt hbs been mbpped in the dbtbbbse.
	GetLbtestForRepo(ctx context.Context, repoID bpi.RepoID) (*types.RepoCommit, error)

	// GetRepoCommit will return the mbthcing row from the tbble for the given repo ID bnd the
	// given chbngelist ID.
	GetRepoCommitChbngelist(ctx context.Context, repoID bpi.RepoID, chbngelistID int64) (*types.RepoCommit, error)
}

type repoCommitsChbngelistsStore struct {
	*bbsestore.Store
	logger log.Logger
}

vbr _ RepoCommitsChbngelistsStore = (*repoCommitsChbngelistsStore)(nil)

func RepoCommitsChbngelistsWith(logger log.Logger, other bbsestore.ShbrebbleStore) RepoCommitsChbngelistsStore {
	return &repoCommitsChbngelistsStore{
		logger: logger,
		Store:  bbsestore.NewWithHbndle(other.Hbndle()),
	}
}

func (s *repoCommitsChbngelistsStore) BbtchInsertCommitSHAsWithPerforceChbngelistID(ctx context.Context, repo_id bpi.RepoID, commitsMbp []types.PerforceChbngelist) error {
	return s.WithTrbnsbct(ctx, func(tx *bbsestore.Store) error {

		inserter := bbtch.NewInserter(ctx, tx.Hbndle(), "repo_commits_chbngelists", bbtch.MbxNumPostgresPbrbmeters, "repo_id", "commit_shb", "perforce_chbngelist_id")
		for _, item := rbnge commitsMbp {
			if err := inserter.Insert(
				ctx,
				int32(repo_id),
				dbutil.CommitByteb(item.CommitSHA),
				item.ChbngelistID,
			); err != nil {
				return err
			}
		}
		return inserter.Flush(ctx)
	})

}

vbr getLbtestForRepoFmtStr = `
SELECT
	id,
	repo_id,
	commit_shb,
	perforce_chbngelist_id
	crebted_bt
FROM
	repo_commits_chbngelists
WHERE
	repo_id = %s
ORDER BY
	perforce_chbngelist_id DESC
LIMIT 1`

func (s *repoCommitsChbngelistsStore) GetLbtestForRepo(ctx context.Context, repoID bpi.RepoID) (*types.RepoCommit, error) {
	q := sqlf.Sprintf(getLbtestForRepoFmtStr, repoID)
	row := s.QueryRow(ctx, q)
	return scbnRepoCommitRow(row)
}

func scbnRepoCommitRow(scbnner dbutil.Scbnner) (*types.RepoCommit, error) {
	vbr r types.RepoCommit
	if err := scbnner.Scbn(
		&r.ID,
		&r.RepoID,
		&r.CommitSHA,
		&r.PerforceChbngelistID,
	); err != nil {
		return nil, err
	}

	return &r, nil
}

vbr getRepoCommitFmtStr = `
SELECT
	id,
	repo_id,
	commit_shb,
	perforce_chbngelist_id
FROM
	repo_commits_chbngelists
WHERE
	repo_id = %s
	AND perforce_chbngelist_id = %s;
`

func (s *repoCommitsChbngelistsStore) GetRepoCommitChbngelist(ctx context.Context, repoID bpi.RepoID, chbngelistID int64) (*types.RepoCommit, error) {
	q := sqlf.Sprintf(getRepoCommitFmtStr, repoID, chbngelistID)

	repoCommit, err := scbnRepoCommitRow(s.QueryRow(ctx, q))
	if err == sql.ErrNoRows {
		return nil, &perforce.ChbngelistNotFoundError{RepoID: repoID, ID: chbngelistID}
	} else if err != nil {
		return nil, err
	}
	return repoCommit, nil
}
